+++ .eslintrc.js
@@ -15,5 +15,6 @@ module.exports = {
     quotes: ["error", "double"],
     curly: ["error", "all"],
     "no-console": ["error", { allow: ["info", "warn", "error"] }],
+    "no-use-before-define": ["error", { functions: false, classes: false }],
   },
 };
